--- src/components/ItemsPage.js
@@ -1,3 +1,0 @@
-function ItemsPage() {
-  return <></>;
-}
